+++ hw7/application/src/lenet/ShapeTy.h
@@ -10,7 +10,7 @@ public:
 
 	ShapeTy(uint n_, uint c_, uint h_, uint w_);
 	ShapeTy();
-	ShapeTy(int64_t* ptr, int num);
+	ShapeTy(int32_t* ptr, int num);
 	~ShapeTy() {}
 
 	// Default: the python-array-like presentation
@@ -22,7 +22,7 @@ public:
 		const std::string suffix="]") const;
 	uint size() const;
 	bool operator==(const ShapeTy& rhs) const;
-	void put(int64_t* ptr, int num);
+	void put(int32_t* ptr, int num);
 	ShapeTy to_strides(int dim) const;
 	inline ShapeTy to_strides() const {return this->to_strides(4);};
 	uint Idx(uint ni, uint ci, uint hi, uint wi,
